requirements.txt
requests
pylint
mypy
astor
networkx
fastapi
uvicorn
python-dotenv
google-generativeai
radon
